[#97] refactor: 컴포넌트 모듈로 각각 분리
