+++ src/components/Detail/ProductReviews/ControlButtons.tsx
@@ -1,0 +1,54 @@
+'use client';
+
+import { ControlButtonsProps } from '@/components/Detail/types';
+import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { deleteReview } from '@/shared/@common/apis';
+import { productKeys } from '@/app/[category]/[product]/queryKeyFactories';
+
+export const ControlButtons = ({
+  accessToken,
+  reviewId,
+  productId,
+  filter,
+}: ControlButtonsProps) => {
+  const currentFilter = filter ?? 'recent';
+
+  const queryClient = useQueryClient();
+  const deleteReviewMutation = useMutation({
+    mutationFn: async () => {
+      await deleteReview(reviewId, accessToken);
+    },
+    onSettled: () => {
+      queryClient.invalidateQueries({
+        queryKey: productKeys.reviews(productId, currentFilter),
+      });
+      queryClient.invalidateQueries({
+        queryKey: productKeys.detail(productId),
+      });
+    },
+  });
+
+  const handleDeleteClick = () => {
+    deleteReviewMutation.mutate();
+  };
+
+  const handleEditClick = () => {
+    alert('리뷰 수정 모달 열림');
+  };
+
+  return (
+    <div className="flex gap-[10px] underline text-gray-9F">
+      <button onClick={handleEditClick} type="button">
+        수정
+      </button>
+      <button
+        onClick={handleDeleteClick}
+        disabled={deleteReviewMutation.isPending}
+        className={`${deleteReviewMutation.isPending ? 'cursor-not-allowed' : 'cursor-pointer'}`}
+        type="button"
+      >
+        삭제
+      </button>
+    </div>
+  );
+};
